Refactor Python script
